Create the Services page
The Services page currently has two sections: one for submitting
service opportunities, and another for viewing current ones.  The
submitting section is complete, linking to a google form, while the
service opportunities section is only partially complete.  It has a
link to the spreadsheet of current service opportunities, but the
embed doesn't work properly.