Read data from google sheet for home page
The data is reread each request, so the home page unfortunately
takes around 2-3 seconds to load properly.  Additionally, the data
is hosted on a spreadsheet under my personal google account
(kleinjason234@gmail.com) until I have access to put it on the non-
FCPS NHS one.  The app depends on the capitalization and spelling
of the headers.