Make needed changes to host on Director
The main area for change is the fact that it isn't hosted
at exactly the origin but instead at a subpath.  That means that
all of the requests need to have "/nhs/" put in front of them.
However, this was partly handled by director itself, so I couldn't
just tell *every* request to have "/nhs/" appended in front of it.
I created a custom `url_for()` function that appends this which
can be toggled depending on development status (as well as a way
for `register.js` to know where the service worker file is
located), but I haven't figured out how to append "/nhs/" to
all of the urls cached by `sw.js` in a way that's sensitive
to development/production status.