Set cache to expire after some time
This will change the way caching works.  By default, it will try
to serve anything that isn't the home page from cache first.  If
that is expired, then it will try to fetch it from the network.
If the network is down or gives a bad response code, then it will
use the (possibly out of date) files from the cache, so that the
PWA works offline.  If the request wasn't found inside of the
cache, then it will effectively pass it through.  This should
allow for new updates to be received by users within at most 2
hours of release.  This should fix issues with persistent cache.